# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Capture API Simulator API",
    "description": "Deterministic traffic-forensics simulator. Browsers cannot call it directly (no CORS): proxy it from your own server. Optional fields are absent rather than null; times are ISO-8601 UTC with milliseconds; session ids are uint64 decimal strings.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "List cases",
        "description": "Newest first. Cases are shared by the whole team; observers may read them.",
        "operationId": "listCases",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor of the next page.",
              "title": "Cursor"
            },
            "description": "Opaque cursor of the next page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseList"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Open a case",
        "operationId": "createCase",
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Createcase"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Read one case",
        "operationId": "getCase",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "cases"
        ],
        "summary": "Update a case",
        "description": "JSON merge-patch over `{title, status, severity, summary, owner_id}`; `summary: null` removes the summary. `If-Match` is REQUIRED (428 without it, 412 `etag_mismatch` with a stale tag). Cases are shared, and the teammate bot edits CASE-0002 every 90 s, so a client that caches an ETag will meet a 412 sooner or later.",
        "operationId": "updateCase",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The case's current ETag, e.g. `\"v7\"`.",
              "examples": [
                "\"v7\""
              ],
              "title": "If-Match"
            },
            "description": "The case's current ETag, e.g. `\"v7\"`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CasePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "412": {
            "description": "Precondition failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Updatecase"
                }
              }
            }
          },
          "428": {
            "description": "Precondition required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/evidence": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Pin a session as evidence",
        "description": "201 with the updated case, or 200 `{\"duplicate\": true}` when that session is already pinned (pinning twice changes nothing, not even the version).",
        "operationId": "pinEvidence",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pinned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "200": {
            "description": "Already pinned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDuplicate"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Pinevidence"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/evidence/{session_id}": {
      "delete": {
        "tags": [
          "cases"
        ],
        "summary": "Unpin a session",
        "description": "Always 204 for an existing case, whether or not that session was pinned.",
        "operationId": "unpinEvidence",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/notes": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Append a note",
        "description": "Notes are append-only; adding one bumps the case version.",
        "operationId": "addNote",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Addnote"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/exports": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Export a case",
        "description": "Starts a background job and answers 202. Poll `GET /v1/exports/{id}` until `state` is `ready`, then download the zip. Sending the same `Idempotency-Key` again returns the same job with `Idempotent-Replayed: true`.",
        "operationId": "createExport",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "8-64 characters of `[A-Za-z0-9_-]`.",
              "title": "Idempotency-Key"
            },
            "description": "8-64 characters of `[A-Za-z0-9_-]`."
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Export"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Createexport"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "summary": {
            "type": "string",
            "maxLength": 2000,
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "severity"
        ],
        "title": "CaseCreate"
      },
      "CaseList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CaseList"
      },
...
      "ErrorBody": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable message."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable snake_case error code."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorBody",
        "description": "Error envelope of every domain error. Extra keys carry machine-readable context."
      },
...
      "EvidenceDuplicate": {
        "properties": {
          "duplicate": {
            "type": "boolean",
            "const": true,
            "title": "Duplicate"
          }
        },
        "type": "object",
        "required": [
          "duplicate"
        ],
        "title": "EvidenceDuplicate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NoteCreate": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "body"
        ],
        "title": "NoteCreate"
      },
...
      "ValidationErrorBody": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorBody",
        "description": "FastAPI's default 422 body: detail is a LIST here, unlike ErrorBody."
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Access token `at_…` from POST /v1/auth/login or /v1/auth/refresh.",
        "scheme": "bearer"
      }
    }
  },
...
}